Add live announcements for carousel selections
