+++ utils.js
@@ -9,3 +9,14 @@ export function adjustColor(col, amt) {
 export function showError(msg) {
   console.error(msg);
 }
+
+export function announce(message) {
+  const announcement = document.createElement('div');
+  announcement.setAttribute('aria-live', 'polite');
+  announcement.setAttribute('aria-atomic', 'true');
+  announcement.style.position = 'absolute';
+  announcement.style.left = '-9999px';
+  announcement.textContent = message;
+  document.body.appendChild(announcement);
+  setTimeout(() => announcement.remove(), 1000);
+}
